app/requirements.txt
flask
streamlit
databricks-sdk
boto3
pytest
matplotlib